+++ tests/tuxemon/test_camera.py
@@ -10,13 +10,10 @@ from tuxemon.camera import (
     SPEED_RIGHT,
     SPEED_UP,
     Camera,
-    CameraInputHandler,
-    CameraManager,
     project,
     unproject,
 )
 from tuxemon.math import Vector2
-from tuxemon.platform.const import intentions
 
 # For the entity at position (5.0, 5.0):
 # The project function calculates:
@@ -123,94 +120,3 @@ class TestCamera(unittest.TestCase):
     def test_move_right(self):
         self.camera.move_right()
         self.assertEqual(self.camera.position.x, 88 + SPEED_RIGHT)
-
-
-class TestCameraManager(unittest.TestCase):
-    def setUp(self):
-        self.manager = CameraManager()
-        self.camera1 = Mock(spec=Camera)
-        self.camera2 = Mock(spec=Camera)
-
-    def test_add_camera(self):
-        self.manager.add_camera(self.camera1)
-        self.assertIn(self.camera1, self.manager.cameras)
-        self.assertEqual(self.manager.active_camera, self.camera1)
-
-    def test_set_active_camera(self):
-        self.manager.add_camera(self.camera1)
-        self.manager.add_camera(self.camera2)
-        self.manager.set_active_camera(self.camera2)
-        self.assertEqual(self.manager.active_camera, self.camera2)
-
-    def test_update(self):
-        self.manager.add_camera(self.camera1)
-        self.manager.update(0.1)
-        self.camera1.update.assert_called_once()
-
-    def test_handle_input(self):
-        self.manager.add_camera(self.camera1)
-        self.camera1.free_roaming_enabled = True
-        event = Mock()
-        self.manager.input_handler.handle_input = Mock()
-        self.manager.handle_input(event)
-        self.manager.input_handler.handle_input.assert_called_once_with(event)
-
-    def test_get_active_camera(self):
-        self.manager.add_camera(self.camera1)
-        self.assertEqual(self.manager.get_active_camera(), self.camera1)
-
-
-class TestCameraInputHandler(unittest.TestCase):
-    def setUp(self):
-        self.camera = Mock(spec=Camera)
-        self.handler = CameraInputHandler(self.camera)
-
-    def test_handle_input_free_roaming_held_up(self):
-        self.camera.free_roaming_enabled = True
-        event = Mock()
-        event.held = True
-        event.pressed = False
-        event.button = intentions.UP
-        self.handler.handle_input(event)
-        self.camera.move_up.assert_called_once()
-
-    def test_handle_input_free_roaming_pressed_down(self):
-        self.camera.free_roaming_enabled = True
-        event = Mock()
-        event.held = False
-        event.pressed = True
-        event.button = intentions.DOWN
-        self.handler.handle_input(event)
-        self.camera.move_down.assert_called_once()
-
-    def test_handle_input_free_roaming_disabled(self):
-        self.camera.free_roaming_enabled = False
-        event = Mock()
-        event.held = True
-        event.button = intentions.UP
-        self.handler.handle_input(event)
-        self.camera.move_up.assert_not_called()
-
-    def test_handle_input_return_event(self):
-        self.camera.free_roaming_enabled = True
-        event = Mock()
-        event.held = True
-        event.button = intentions.UP
-        returned_event = self.handler.handle_input(event)
-        self.assertEqual(event, returned_event)
-
-    def test_handle_input_left(self):
-        self.camera.free_roaming_enabled = True
-        event = Mock()
-        event.held = True
-        event.button = intentions.LEFT
-        self.handler.handle_input(event)
-        self.camera.move_left.assert_called_once()
-
-    def test_handle_input_right(self):
-        self.camera.free_roaming_enabled = True
-        event = Mock()
-        event.held = True
-        event.button = intentions.RIGHT
-        self.handler.handle_input(event)
-        self.camera.move_right.assert_called_once()
