+++ tests/tuxemon/test_sprite_renderer.py
